Fix non-deterministic tests
Many of these tests were reliant on the test runner
environment having a specific local time zone set.
I updated the tests to set their own time zone, and then
fixed the failing tests.

One test also failed because it assumed the current date
was in the year 2018! I provided a default date into that
one, which fixed that.